requirements.txt
flask
tweepy
textblob
gunicorn
python-dotenv
